+++ mobile-app/src/components/notification-components/FriendRequest.tsx
@@ -3,8 +3,8 @@ import {
   deleteNotificationAtom,
   friendNotificationAtom,
   getUserInfoAtom,
+  numberOfMutualFriendsAtom,
 } from "@/src/atoms/atoms";
-import { numberOfMutualFriendsAtom } from "@/src/atoms/friendsAtom";
 import { useAtomValue } from "jotai";
 import { NotifProfilePicture } from "../ProfilePicture";
 import { Text, View } from "../Themed";
